fix 9/25 80%

+++ ft_strdup.c
@@ -6,7 +6,7 @@
 /*   By: kdodo <kdodo@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/24 15:21:45 by kdodo             #+#    #+#             */
-/*   Updated: 2023/09/25 16:27:37 by kdodo            ###   ########.fr       */
+/*   Updated: 2023/09/25 17:53:42 by kdodo            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_strdup(const char *s1)
 {
-	int		len;
-	int		i;
+	size_t	len;
+	size_t	i;
 	char	*str;
 
 	len = ft_strlen(s1);
